Moved the syscall arg display implementations to a new module
When supporting new types in the syscall logger, the `SyscallDataDisplay`,
`SyscallPtrDisplay`, and `TryFromSyscallReg` traits should be implemented in
`syscall::type_formatting`.
