+++ src/main/host/syscall/mod.rs
@@ -3,6 +3,7 @@ use crate::host::descriptor::{File, FileState};
 
 pub mod formatter;
 pub mod handler;
+pub mod type_formatting;
 
 // The helpers defined here are syscall-related but not handler-specific.
